# SistemaAlertasBackEnd/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SistemaAlertasBackEnd, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/alertas/alertas": {
      "get": {
        "tags": [
          "AlertaEndpoint"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetAllAlertasDTO"
              }
            }
          }
        }
      }
    },
    "/alertas/alertas/{id}": {
      "get": {
        "tags": [
          "AlertaEndpoint"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetAllAlertasDTO"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lecturas/lecturas": {
      "post": {
        "tags": [
          "LecturaEndPoint"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CrearLecturaDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/GetAllLecturasDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          }
        }
      },
      "get": {
        "tags": [
          "LecturaEndPoint"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetAllLecturasDTO"
              }
            }
          }
        }
      }
    },
    "/lecturas/lecturas/{id}": {
      "get": {
        "tags": [
          "LecturaEndPoint"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetAllLecturasDTO"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/sensores": {
      "post": {
        "tags": [
          "SensorEndpoint"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CrearSensorDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/GetAllSensoresDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SensorEndpoint"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetAllSensoresDTO"
              }
            }
          }
        }
      }
    },
    "/api/sensores/{id}": {
      "get": {
        "tags": [
          "SensorEndpoint"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetAllSensoresDTO"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "SensorEndpoint"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CrearSensorDTO"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "SensorEndpoint"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/usuarios/registrar": {
      "post": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioCrearDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RespuestaAutenticacionDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IdentityError"
              }
            }
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioCrearDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RespuestaAutenticacionDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/usuarios/usuarios": {
      "get": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/usuarios/usuarios/{id}": {
      "get": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioDTO"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ActualizarUsuarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/usuarios/asignar-rol": {
      "post": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditarClaimDTO"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IdentityError"
              }
            }
          }
        }
      }
    },
    "/usuarios/remover-rol-admin": {
      "post": {
        "tags": [
          "UsuariosEndPoints"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditarClaimDTO"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ActualizarUsuarioDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "correoElectronico": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CrearLecturaDTO": {
      "type": "object",
      "properties": {
        "sensorId": {
          "format": "int32",
          "type": "integer"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "unity": {
          "type": "string"
        },
        "ph_parameter": {
          "format": "double",
          "type": "number"
        },
        "orp_parameter": {
          "format": "double",
          "type": "number"
        },
        "turbidez_parameter": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "CrearSensorDTO": {
      "type": "object",
      "properties": {
        "nombreSensor": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "EditarClaimDTO": {
      "type": "object",
      "properties": {
        "correoElectronico": {
          "type": "string"
        },
        "tipoClaim": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "GetAllAlertasDTO": {
      "type": "object",
      "properties": {
        "alertaId": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "level": {
          "type": "string"
        },
        "sensorId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "GetAllLecturasDTO": {
      "type": "object",
      "properties": {
        "readId": {
          "format": "int32",
          "type": "integer"
        },
        "sensorId": {
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "unity": {
          "type": "string"
        },
        "ph_parameter": {
          "format": "double",
          "type": "number"
        },
        "orp_parameter": {
          "format": "double",
          "type": "number"
        },
        "turbidez_parameter": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "GetAllSensoresDTO": {
      "type": "object",
      "properties": {
        "sensorId": {
          "format": "int32",
          "type": "integer"
        },
        "nombreSensor": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "lecturaEntidades": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetAllLecturasDTO"
          }
        }
      },
      "additionalProperties": false
    },
    "HttpValidationProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "errors": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "additionalProperties": { }
    },
    "IdentityError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "RespuestaAutenticacionDTO": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "expiracion": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UsuarioCrearDTO": {
      "type": "object",
      "properties": {
        "correoElectronico": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UsuarioDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "correoElectronico": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "derechosUsuario": {
          "type": "string"
        },
        "fechaRegistro": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}